Telegram Bot API August 15, 2025 updates (v9.2)

+++ telegramium-core/src/main/scala/telegramium/bots/AffiliateInfo.scala
@@ -18,7 +18,7 @@ package telegramium.bots
   */
 final case class AffiliateInfo(
   commissionPerMille: Int,
-  amount: Int,
+  amount: Long,
   affiliateUser: Option[User] = Option.empty,
   affiliateChat: Option[Chat] = Option.empty,
   nanostarAmount: Option[Int] = Option.empty
