+++ telegramium-core/src/main/scala/telegramium/bots/LabeledPrice.scala
@@ -10,4 +10,4 @@ package telegramium.bots
   * 1.45 pass amount = 145. See the exp parameter in currencies.json, it shows the number of digits past the decimal
   * point for each currency (2 for the majority of currencies).
   */
-final case class LabeledPrice(label: String, amount: Int)
+final case class LabeledPrice(label: String, amount: Long)
