+++ telegramium-core/src/main/scala/telegramium/bots/StarAmount.scala
@@ -8,4 +8,4 @@ package telegramium.bots
   *   Optional. The number of 1/1000000000 shares of Telegram Stars; from -999999999 to 999999999; can be negative if
   *   and only if amount is non-positive
   */
-final case class StarAmount(amount: Int, nanostarAmount: Option[Int] = Option.empty)
+final case class StarAmount(amount: Long, nanostarAmount: Option[Int] = Option.empty)
